+++ src/components/view/CalendarView.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types"
-import { useState } from "react"
+import { useEffect, useRef, useState } from "react"
 import supabase from "../../libs/supabase"
 import { formatToYYMMDD } from "../../utils/date"
 import FixedBottomButton from "../button/FixedBottomButton"
@@ -21,20 +21,85 @@ CalendarView.propTypes = {
 }
 
 export default function CalendarView({ eventInfo, participantId }) {
-  const [availabilityInfo, setAvailabilityInfo] = useState({
-    eventId: eventInfo.id,
-    participantId: participantId,
-    availableTimeslots: "",
-  })
   const [selectedSlots, setSelectedSlots] = useState(new Set())
+  const isInitialLoadTimeSlot = useRef(true)
 
+  // 기존에 저장한 시간 조회 & 상태에 저장
+  const initTimeSlot = async () => {
+    try {
+      const { data } = await supabase
+        .from("availabilities")
+        .select("available_timeslots")
+        .eq("event_id", eventInfo?.id)
+        .eq("participant_id", participantId)
+
+      setSelectedSlots(new Set(data[0]?.available_timeslots))
+    } catch (err) {
+      console.error("Supabase Insert Error:", err)
+    } finally {
+      isInitialLoadTimeSlot.current = false
+    }
+  }
+
+  useEffect(() => {
+    initTimeSlot()
+  }, [])
+
+  // 선택한 시간 저장
+  const saveTimeSlot = async (newSlots) => {
+    try {
+      // 중복값 체크
+      const { data: existAvailability } = await supabase
+        .from("availabilities")
+        .select("id")
+        .eq("event_id", eventInfo?.id)
+        .eq("participant_id", participantId)
+
+      if (existAvailability.length) {
+        // 기존 timeslots 업데이트
+        const { data, error } = await supabase
+          .from("availabilities")
+          .update({ available_timeslots: [...newSlots] })
+          .eq("id", existAvailability[0].id)
+          .select()
+          .single()
+        console.log("---기존 timeslots 업데이트---")
+        console.log(data)
+        console.log(error)
+      } else {
+        // 신규 입력
+        const { data, error } = await supabase
+          .from("availabilities")
+          .insert({
+            event_id: eventInfo?.id,
+            participant_id: participantId,
+            available_timeslots: [...newSlots],
+          })
+          .select()
+          .single()
+        console.log("---신규 입력---")
+        console.log(data)
+        console.log(error)
+      }
+    } catch (err) {
+      console.error("Supabase Insert Error:", err)
+    }
+  }
+
+  useEffect(() => {
+    if (!isInitialLoadTimeSlot.current && selectedSlots.size > 0) {
+      saveTimeSlot(selectedSlots)
+    }
+  }, [selectedSlots])
+
+  // 선택한 시간 초기화
   const resetTimeSlot = async () => {
     try {
       await supabase
         .from("availabilities")
         .delete()
-        .eq("event_id", availabilityInfo?.eventId)
-        .eq("participant_id", availabilityInfo?.participantId)
+        .eq("event_id", eventInfo?.id)
+        .eq("participant_id", participantId)
       setSelectedSlots(new Set())
     } catch (err) {
       console.error("Supabase Insert Error:", err)
@@ -80,7 +145,6 @@ export default function CalendarView({ eventInfo, participantId }) {
               selectedSlots={selectedSlots}
               setSelectedSlots={setSelectedSlots}
               timegridInfo={timegridInfo}
-              availabilityInfo={availabilityInfo}
             />
           </section>
